refactor: Migrate Paint in a Gradle module.
- Create colorpicker component module.
- Create paint demo app module.

Issue #61
