+++ demos/paint/build.gradle.kts
@@ -1,0 +1,16 @@
+import com.guru.composecookbook.build.dependencies.*
+
+plugins {
+    /**
+     * See [common-compose-module-configs-script-plugin.gradle.kts] file
+     */
+    id("common-compose-module-configs-script-plugin")
+}
+
+dependencies {
+    implementation(project(":theme"))
+    implementation(project(":components:colorpicker"))
+
+    addComposeOfficialDependencies()
+    addCoreAndroidUiDependencies()
+}
