+++ demos/paint/src/main/java/com/guru/composecookbook/paint/PaintActivity.kt
@@ -1,0 +1,38 @@
+package com.guru.composecookbook.paint
+
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.core.content.ContextCompat
+import com.guru.composecookbook.theme.ComposeCookBookTheme
+
+class PaintActivity : ComponentActivity() {
+    @ExperimentalAnimationApi
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        window.statusBarColor = ContextCompat.getColor(this, android.R.color.transparent)
+        setContent {
+            ComposeCookBookTheme {
+                PaintApp()
+            }
+        }
+    }
+
+    companion object {
+        fun newIntent(context: Context) = Intent(context, PaintActivity::class.java)
+    }
+}
+
+@ExperimentalAnimationApi
+@Preview(showBackground = true)
+@Composable
+fun DefaultPreview4() {
+    ComposeCookBookTheme {
+        PaintApp()
+    }
+}
