+++ settings.gradle.kts
@@ -16,3 +16,5 @@ include(":demos:twitter")
 include(":demos:youtube")
 include(":demos:gmail")
 include(":components:fab")
+include(":demos:paint")
+include(":components:colorpicker")
